feat: update dependencies and refactor settings components
