+++ apps/api/package.json
@@ -21,28 +21,28 @@
     "dependencies": {
         "@nestjs-modules/mailer": "^2.0.2",
         "@nestjs/axios": "^4.0.1",
-        "@nestjs/common": "^11.1.7",
-        "@nestjs/core": "^11.1.7",
-        "@nestjs/event-emitter": "^3.0.1",
+        "@nestjs/common": "^11.1.8",
         "@nestjs/config": "^4.0.2",
-        "@nestjs/jwt": "^11.0.0",
+        "@nestjs/core": "^11.1.8",
+        "@nestjs/event-emitter": "^3.0.1",
+        "@nestjs/jwt": "^11.0.1",
         "@nestjs/passport": "^11.0.5",
-        "@nestjs/platform-express": "^11.1.3",
-        "@nestjs/schedule": "^6.0.0",
+        "@nestjs/platform-express": "^11.1.8",
+        "@nestjs/schedule": "^6.0.1",
         "@nestjs/throttler": "^6.4.0",
         "@packages/agent": "workspace:*",
         "@packages/monitoring": "workspace:*",
-        "@vercel/sdk": "^1.11.4",
-        "axios": "^1.12.2",
+        "@vercel/sdk": "^1.17.0",
+        "axios": "^1.13.2",
         "bcrypt": "^6.0.0",
-        "better-sqlite3": "^12.2.0",
+        "better-sqlite3": "^12.4.1",
         "class-transformer": "^0.5.1",
         "class-validator": "^0.14.2",
-        "dotenv": "^17.0.1",
+        "dotenv": "^17.2.3",
         "handlebars": "^4.7.8",
         "helmet": "^8.1.0",
-        "mysql2": "^3.14.4",
-        "nodemailer": "^7.0.7",
+        "mysql2": "^3.15.3",
+        "nodemailer": "^7.0.10",
         "passport": "^0.7.0",
         "passport-github2": "^0.1.12",
         "passport-google-oauth20": "^2.0.0",
@@ -53,24 +53,24 @@
         "rxjs": "^7.8.2"
     },
     "devDependencies": {
-        "@nestjs/cli": "^11.0.7",
-        "@nestjs/schematics": "^11.0.5",
-        "@nestjs/testing": "^11.1.3",
+        "@nestjs/cli": "^11.0.10",
+        "@nestjs/schematics": "^11.0.9",
+        "@nestjs/testing": "^11.1.8",
         "@swc/cli": "^0.6.0",
-        "@swc/core": "^1.12.9",
+        "@swc/core": "^1.15.1",
         "@types/bcrypt": "^6.0.0",
-        "@types/express": "^5.0.3",
+        "@types/express": "^5.0.5",
         "@types/jest": "^29.5.14",
-        "@types/node": "^22.16.0",
+        "@types/node": "^22.19.0",
         "@types/passport-jwt": "^4.0.1",
         "@types/passport-local": "^1.0.38",
         "@types/supertest": "^6.0.3",
-        "globals": "^16.3.0",
+        "globals": "^16.5.0",
         "source-map-support": "^0.5.21",
-        "supertest": "^7.1.3",
-        "ts-loader": "^9.5.2",
+        "supertest": "^7.1.4",
+        "ts-loader": "^9.5.4",
         "ts-node": "^10.9.2",
         "tsconfig-paths": "^4.2.0",
-        "typescript": "^5.8.3"
+        "typescript": "^5.9.3"
     }
 }
