+++ apps/web/package.json
@@ -17,20 +17,20 @@
         "jwt-decode": "^4.0.0",
         "lucide-react": "^0.542.0",
         "next": "16.0.1",
-        "next-intl": "^4.4.0",
+        "next-intl": "^4.5.0",
         "path-to-regexp": "^8.3.0",
         "react": "19.1.0",
         "react-dom": "19.1.0",
         "sonner": "^2.0.7",
-        "tailwind-merge": "^3.3.1",
+        "tailwind-merge": "^3.4.0",
         "zod": "^3.25.76"
     },
     "devDependencies": {
-        "@tailwindcss/postcss": "^4.1.16",
-        "@types/node": "^20.19.23",
+        "@tailwindcss/postcss": "^4.1.17",
+        "@types/node": "^20.19.24",
         "@types/react": "^19.2.2",
         "@types/react-dom": "^19.2.2",
-        "tailwindcss": "^4.1.16",
+        "tailwindcss": "^4.1.17",
         "typescript": "^5.9.3"
     }
 }
