+++ packages/agent/package.json
@@ -103,74 +103,74 @@
     },
     "devDependencies": {
         "@eslint/eslintrc": "^3.3.1",
-        "@eslint/js": "^9.30.1",
-        "@nestjs/cli": "^11.0.7",
-        "@nestjs/schematics": "^11.0.5",
-        "@nestjs/testing": "^11.1.3",
+        "@eslint/js": "^9.39.1",
+        "@nestjs/cli": "^11.0.10",
+        "@nestjs/schematics": "^11.0.9",
+        "@nestjs/testing": "^11.1.8",
         "@swc/cli": "^0.6.0",
-        "@swc/core": "^1.12.9",
+        "@swc/core": "^1.15.1",
         "@trigger.dev/build": "^4.0.6",
-        "@types/express": "^5.0.3",
+        "@types/express": "^5.0.5",
         "@types/jest": "^29.5.14",
-        "@types/node": "^22.16.0",
+        "@types/node": "^22.19.0",
         "@types/supertest": "^6.0.3",
-        "concurrently": "^9.2.0",
-        "eslint": "^9.30.1",
-        "eslint-config-prettier": "^10.1.5",
-        "eslint-plugin-prettier": "^5.5.1",
-        "globals": "^16.3.0",
+        "concurrently": "^9.2.1",
+        "eslint": "^9.39.1",
+        "eslint-config-prettier": "^10.1.8",
+        "eslint-plugin-prettier": "^5.5.4",
+        "globals": "^16.5.0",
         "jest": "^29.7.0",
         "prettier": "^3.6.2",
         "source-map-support": "^0.5.21",
-        "supertest": "^7.1.1",
-        "ts-jest": "^29.4.0",
-        "ts-loader": "^9.5.2",
+        "supertest": "^7.1.4",
+        "ts-jest": "^29.4.5",
+        "ts-loader": "^9.5.4",
         "ts-node": "^10.9.2",
         "tsconfig-paths": "^4.2.0",
-        "typescript": "^5.8.3",
-        "typescript-eslint": "^8.35.1"
+        "typescript": "^5.9.3",
+        "typescript-eslint": "^8.46.3"
     },
     "dependencies": {
-        "@langchain/anthropic": "^0.3.24",
-        "@langchain/core": "^0.3.62",
+        "@langchain/anthropic": "^0.3.33",
+        "@langchain/core": "^0.3.79",
         "@langchain/deepseek": "^0.0.1",
-        "@langchain/groq": "^0.2.3",
+        "@langchain/groq": "^0.2.4",
         "@langchain/langgraph": "^0.2.74",
-        "@langchain/mistralai": "^0.2.1",
+        "@langchain/mistralai": "^0.2.3",
         "@langchain/openai": "0.5.12",
         "@langchain/textsplitters": "^0.1.0",
         "@nestjs/cache-manager": "^3.0.1",
         "@nestjs/config": "^4.0.2",
         "@nestjs/typeorm": "^11.0.0",
         "@nestjs/websockets": "^11.1.8",
-        "@supabase/supabase-js": "^2.50.3",
+        "@supabase/supabase-js": "^2.80.0",
         "@tavily/core": "^0.3.7",
         "@trigger.dev/sdk": "^4.0.6",
         "@types/cheerio": "^1.0.0",
         "@types/fs-extra": "^11.0.4",
         "@types/string-similarity": "^4.0.2",
-        "@vercel/sdk": "^1.14.0",
-        "axios": "^1.12.2",
-        "bullmq": "^5.56.1",
-        "cache-manager": "^7.2.0",
-        "cheerio": "^1.1.0",
+        "@vercel/sdk": "^1.17.0",
+        "axios": "^1.13.2",
+        "bullmq": "^5.63.0",
+        "cache-manager": "^7.2.4",
+        "cheerio": "^1.1.2",
         "class-transformer": "^0.5.1",
         "class-validator": "^0.14.2",
         "date-fns": "^4.1.0",
         "deepmerge": "^4.3.1",
         "google-sr": "^5.0.0",
         "import-in-the-middle": "^2.0.0",
-        "isomorphic-git": "^1.32.1",
-        "langchain": "^0.3.29",
+        "isomorphic-git": "^1.35.0",
+        "langchain": "^0.3.36",
         "libsodium-wrappers": "^0.7.15",
         "octokit": "^3.2.2",
         "remark-stringify": "^11.0.0",
         "slugify": "^1.6.6",
         "string-similarity": "^4.0.4",
-        "turndown": "^7.2.0",
+        "turndown": "^7.2.2",
         "typeorm": "^0.3.27",
         "unified": "^11.0.5",
-        "yaml": "^2.8.0",
-        "zod": "^3.25.73"
+        "yaml": "^2.8.1",
+        "zod": "^3.25.76"
     }
 }
