+++ packages/cli-shared/package.json
@@ -17,13 +17,13 @@
 	},
 	"dependencies": {
 		"chalk": "^4.1.2",
-		"inquirer": "^12.7.0",
-		"fs-extra": "^11.3.0"
+		"fs-extra": "^11.3.2",
+		"inquirer": "^12.11.0"
 	},
 	"devDependencies": {
-		"@types/inquirer": "^9.0.8",
 		"@types/fs-extra": "^11.0.4",
-		"@types/node": "^22.16.3",
-		"typescript": "^5.8.3"
+		"@types/inquirer": "^9.0.9",
+		"@types/node": "^22.19.0",
+		"typescript": "^5.9.3"
 	}
 }
