+++ packages/monitoring/package.json
@@ -48,28 +48,28 @@
         "rxjs": "^7.8"
     },
     "dependencies": {
-        "@sentry/nestjs": "^10.17.0",
-        "@sentry/profiling-node": "^10.17.0",
-        "posthog-node": "^5.9.2"
+        "@sentry/nestjs": "^10.23.0",
+        "@sentry/profiling-node": "^10.23.0",
+        "posthog-node": "^5.11.2"
     },
     "devDependencies": {
-        "@nestjs/cli": "^11.0.7",
-        "@nestjs/schematics": "^11.0.5",
-        "@nestjs/testing": "^11.1.3",
+        "@nestjs/cli": "^11.0.10",
+        "@nestjs/schematics": "^11.0.9",
+        "@nestjs/testing": "^11.1.8",
         "@swc/cli": "^0.6.0",
-        "@swc/core": "^1.12.9",
-        "@types/express": "^5.0.3",
+        "@swc/core": "^1.15.1",
+        "@types/express": "^5.0.5",
         "@types/jest": "^29.5.14",
-        "@types/node": "^22.16.0",
+        "@types/node": "^22.19.0",
         "@types/supertest": "^6.0.3",
-        "concurrently": "^9.2.0",
+        "concurrently": "^9.2.1",
         "jest": "^29.7.0",
         "source-map-support": "^0.5.21",
-        "supertest": "^7.1.1",
-        "ts-jest": "^29.4.0",
-        "ts-loader": "^9.5.2",
+        "supertest": "^7.1.4",
+        "ts-jest": "^29.4.5",
+        "ts-loader": "^9.5.4",
         "ts-node": "^10.9.2",
         "tsconfig-paths": "^4.2.0",
-        "typescript": "^5.8.3"
+        "typescript": "^5.9.3"
     }
 }
